Split overview components into separate files
